Sources/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Kirill Novgorodtsev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
djoser = "^2.1.0"
django-rest-swagger = "^2.2.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
black = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
